requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
PyMySQL
flask-jwt-extended
flask_marshmallow
flask_migrate
requests
marshmallow-sqlalchemy